+++ components/site/src/lib.rs
@@ -218,7 +218,6 @@ impl Site {
 
         self.tera.register_global_function("get_page", global_fns::make_get_page(&self.pages));
         self.tera.register_global_function("get_section", global_fns::make_get_section(&self.sections));
-        self.tera.register_global_function("get_static_url", global_fns::make_get_static_url(self.config.clone()));
         self.register_get_url_fn();
 
         Ok(())
@@ -226,7 +225,10 @@ impl Site {
 
     /// Separate fn as it can be called in the serve command
     pub fn register_get_url_fn(&mut self) {
-        self.tera.register_global_function("get_url", global_fns::make_get_url(self.permalinks.clone()));
+        self.tera.register_global_function(
+            "get_url",
+            global_fns::make_get_url(self.permalinks.clone(), self.config.clone())
+        );
     }
 
     /// Add a page to the site