+++ CHANGELOG.md
@@ -6,7 +6,8 @@
 root page
 - Make `title` in config optional
 - Improved `gutenberg init` UX and users first experience
-- Add a `get_static_url` global function that can do cachebusting
+- Make `get_url` work for any path with optional cachebusting. Deprecates `link` param
+in favour of `path` to be consistent
 
 ## 0.1.1 (2017-07-16)